requirements/base.txt
# Django
django

# REST
djangorestframework
rest-framework-generic-relations
drf-nested-routers
django-filter
drf-extensions
drf-yasg